tests/wpt/tests/css/css-backgrounds/parsing/background-attachment-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().backgroundAttachment</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-attachment">
<meta name="assert" content="background-attachment computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-attachment", "fixed");

// See background-computed.html for a test with multiple background images.
test_computed_value("background-attachment", "scroll, fixed, local", "scroll");
</script>
</body>
</html>
